exam-backend-fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.10.6
python-dotenv==1.0.1
requests==2.32.3
pypdf==5.4.0
python-multipart==0.0.20
